feat: comment feature

+++ src/controllers/postController.ts
@@ -3,6 +3,7 @@ import { PostService } from "@/services/postService";
 import { Arrays } from "@/utils";
 import { Files } from "@/utils/Files";
 import { ResponseBody } from "@/utils/ResponseBody";
+import { PostValidator } from "@/validators/PostValidator";
 import { configDotenv } from "dotenv";
 import { Request, Response } from "express";
 
@@ -47,37 +48,34 @@ class PostController {
         }
     }
 
-    createPost = async (req: Request, res: Response) => {
-        const invalidFields = ResponseBody.getInvalidPostFields(req.body);
+    create = async (req: Request, res: Response) => {
         const { title, content, userId, categoryId } = req.body;
-        if(invalidFields.length > 0) {
-            return res.status(400).json({
-                message: 'Invalid fields, these fields are required',
-                fields: invalidFields
-            });
+        const validator = new PostValidator(req);
+        if(!validator.validate()) {
+            return res.status(400).json(validator.getError());
         }
-
+        
         const file = req.file;
         const thumbnail = Files.getPublicPath(file);
-
         const post = Post.build({ title, content, userId, categoryId, thumbnail });
-        const result = await this.postService.create(post);
-        if(result) {
-            res.status(201).json({
+        try {
+            const result = await this.postService.create(post);
+            res.json({
                 statusCode: 201,
                 message: 'Post created successfully',
                 data: result
             });
-        } else {
+        } catch (error) {
             res.status(500).json({ 
                 statusCode: 500,
-                message: 'Internal server error' 
+                message: 'Internal server error',
+                error: error,
             });
         }
     }
 
-    updatePost = async (req: Request, res: Response) => {
-        const { title, content, userId, categoryId } = req.body;
+    update = async (req: Request, res: Response) => {
+        const { title, content, userId, categoryId, isUpload } = req.body;
         const id = parseInt(req.params.id);
         const invalidFields = ResponseBody.getInvalidPostFields(req.body);
         if(invalidFields.length > 0) {
@@ -88,12 +86,14 @@ class PostController {
             });
         }
         const file = req.file;
-        const thumbnail = Files.getPublicPath(file);
+        const oldPost = await this.postService.findById(id);
+        const thumbnail = (isUpload === 'true') ? Files.getPublicPath(file) : oldPost?.thumbnail;
         const post = Post.build({ title, content, userId, categoryId, thumbnail });
         try {
             const oldPost = await this.postService.findById(id);
             const entryDir = process.env.UPLOAD_ENTRY_DIR as string;
-            Files.removeSyncByPath(`${entryDir}${oldPost?.thumbnail}`);
+            if(isUpload === 'true' && oldPost?.thumbnail)
+                Files.removeSyncByPath(`${entryDir}${oldPost?.thumbnail}`);
             const result = await this.postService.update(id, post);
             const [affectedCount] = result;
             if(affectedCount > 0) {
@@ -122,23 +122,31 @@ class PostController {
 
     deletePost = async (req: Request, res: Response) => {
         const id = parseInt(req.params.id);
-        const affectedCount = await this.postService.delete(id);
-        const post = await this.postService.findById(id);
-        if(post) {
-            const entryDir = process.env.UPLOAD_ENTRY_DIR as string;
-            Files.removeSyncByPath(`${entryDir}${post.thumbnail}`);
-        }
-        if(affectedCount > 0) {
-            res.json({
-                statusCode: 200,
-                message: 'Post deleted successfully',
-                affectedCount: affectedCount
-            });
-        } 
-        else {
-            res.status(404).json({
-                statusCode: 404,
-                message: 'Post not found' 
+        try {
+            const affectedCount = await this.postService.delete(id);
+            const post = await this.postService.findById(id);
+            if(post) {
+                const entryDir = process.env.UPLOAD_ENTRY_DIR as string;
+                Files.removeSyncByPath(`${entryDir}${post.thumbnail}`);
+            }
+            if(affectedCount > 0) {
+                res.json({
+                    statusCode: 200,
+                    message: 'Post deleted successfully',
+                    affectedCount: affectedCount
+                });
+            } 
+            else {
+                res.status(404).json({
+                    statusCode: 404,
+                    message: 'Post not found' 
+                });
+            }
+        } catch (error) {
+            res.status(500).json({ 
+                statusCode: 500,
+                message: 'Internal server error',
+                error: error,
             });
         }
     }
